new object added and assimp2bin improvements
quad object added (screen aligned too)
assimp2bin now export vertices colors

+++ src/renderer/renderer.hpp
@@ -64,7 +64,7 @@
 			public:
 				void setViewport(GLsizei w, GLsizei h) {
 					if (currCamera) {
-						currCamera->aspect = w / h;
+						currCamera->aspect = w / (float) h;
 						currCamera->buildProjection();
 					}
